Created two JSP pages and a ColorUtils.java class file.
ColorUtils.java class file is used to create a random color by calling
the method in the JSP file.

ex01-a.jsp set the bg color to a random Color whenever the page was
refresed. 

ec01-b.jsp set the bg color to a specific color specified by the client
by adding ?bg=color to the end of the URL path, where color is any color
the client wants to see his/her bg in. Client/usr must enter to see the
bg color change